Removed Quest Saving
Commented out code for quest saving so there are still zombie code among the scripts.

Next agenda is to change level complete behavior to either go directly to the next level or go back to level select.

Idea:
Once level is done prompt the player whether to continue to the next level or go back to the level select screen